requirements.txt
Django==1.11
django-filter==1.1.0
djangorestframework==3.7.7
dj-database-url==0.4.2
dj-static==0.0.6
python-decouple==3.1
static3==0.7.0
gunicorn==19.4.1
django-extensions==1.9.8
django-test-without-migrations==0.6
django-cors-headers==2.1.0
mysql-connector-python-rf
mysqlclient
psycopg2
